+++ examples/minimum/locals.tf
@@ -1,0 +1,3 @@
+locals {
+  tags = merge({ "Name" = var.name }, var.tags)
+}