+++ examples/minimum/main.tf
@@ -1,5 +1,5 @@
 module "minimum" {
   source = "./../.."
   name   = var.name
-  scope  = var.scope
+  tags   = local.tags
 }