+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix: CKV_AWS_192 #"Ensure WAF prevents message lookup in Log4j2. See CVE-2021-44228 aka log4jshell"
 - fix: CKV2_AWS_31 #"Ensure WAF2 has a Logging Configuration"
 - feat: Add remaining statement features for the rule block
+- feat: Add an example showcasing Global scope(cloudfront)
 - feat: Add more examples on different usage of the module showcasing the usage of most dynamic blocks
 - feat: Add classic waf submodule and examples
 - feat: Add rule group features